Should not write empty data to file
